@@ -5,7 +5,7 @@
     <AssemblyName>DSharpPlus.Test</AssemblyName>
     <RootNamespace>DSharpPlus.Test</RootNamespace>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <VersionPrefix>4.0.0</VersionPrefix>
     <LangVersion>7.2</LangVersion>
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
